refactor: rename entity files to camelCase
